imx8mn/env: Try to boot from different partitions
With the upcoming A/B boot we are also introducing a new partition
scheme which in return will no longer boot with the hardcoded mmcpart=1
which is used for the variscite debian boot process.
To be able to recover in case of a broken bootscript we now scan both
root_a and root_b paritions for bootable images as well as the legacy partition
one to keep compatibility with the variscite images.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>